Extracted document upload from view
